TheBtcTracker (3)/TheBtcTracker/requirements.txt.txt
requests
python-telegram-bot==20.4
python-dotenv